Changes to database: (i) Ids as long type and (ii) decimal / BigDecimal to represent prices, commission, transaction value, ...

+++ src/main/java/com/nefentus/api/entities/Affiliate.java
@@ -1,5 +1,6 @@
 package com.nefentus.api.entities;
 
+import java.math.BigDecimal;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -14,21 +15,21 @@ import java.sql.Timestamp;
 @Table(name = "aff_affiliate")
 public class Affiliate {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "aff_id")
-    private Integer id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "aff_id")
+	private Long id;
 
-    @Column(name = "aff_affiliate_link")
-    private String affiliateLink;
+	@Column(name = "aff_affiliate_link")
+	private String affiliateLink;
 
-    @Column(name = "aff_commission_rate")
-    private Float commissionRate;
+	@Column(name = "aff_commission_rate")
+	private BigDecimal commissionRate;
 
-    @Column(name = "aff_created_at")
-    private Timestamp createdAt;
+	@Column(name = "aff_created_at")
+	private Timestamp createdAt;
 
-    @ManyToOne
-    @JoinColumn(name = "aff_user_id", referencedColumnName = "id")
-    private User user;
+	@ManyToOne
+	@JoinColumn(name = "aff_user_id", referencedColumnName = "id")
+	private User user;
 }
