+++ src/main/java/com/nefentus/api/entities/Hierarchy.java
@@ -1,5 +1,6 @@
 package com.nefentus.api.entities;
 
+import java.math.BigDecimal;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -10,7 +11,7 @@ import java.sql.Timestamp;
 
 @Entity
 @Table(name = "Hierarchy", uniqueConstraints = {
-        @UniqueConstraint(columnNames = {"child_id"})
+		@UniqueConstraint(columnNames = { "child_id" })
 })
 @Getter
 @Setter
@@ -18,26 +19,25 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 public class Hierarchy {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "parent_id")
-    private User parent;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "parent_id")
+	private User parent;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "child_id")
-    private User child;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "child_id")
+	private User child;
 
-    @Column(name = "relationship_type")
-    @Enumerated(EnumType.STRING)
-    private RelationshipType relationshipType;
+	@Column(name = "relationship_type")
+	@Enumerated(EnumType.STRING)
+	private RelationshipType relationshipType;
 
-    @Column(name = "commission_rate")
-    private float commissionRate;
+	@Column(name = "commission_rate")
+	private BigDecimal commissionRate;
 
-    @Column(name = "created_at")
-    private Timestamp createdAt;
+	@Column(name = "created_at")
+	private Timestamp createdAt;
 }
-
