+++ src/main/java/com/nefentus/api/entities/Provision.java
@@ -1,5 +1,6 @@
 package com.nefentus.api.entities;
 
+import java.math.BigDecimal;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,28 +13,28 @@ import java.sql.Timestamp;
 @Table(name = "prv_provision")
 public class Provision {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "prv_id")
-    private Integer id;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "prv_id")
+	private Long id;
 
-    @Column(name = "prv_affiliate_id")
-    private Integer affiliateId;
+	@Column(name = "prv_affiliate_id")
+	private Long affiliateId;
 
-    @Column(name = "prv_transaction_id")
-    private Long transactionId;
+	@Column(name = "prv_transaction_id")
+	private Long transactionId;
 
-    @Column(name = "prv_commission_amount")
-    private Float commissionAmount;
+	@Column(name = "prv_commission_amount")
+	private BigDecimal commissionAmount;
 
-    @Column(name = "prv_created_at")
-    private Timestamp createdAt;
+	@Column(name = "prv_created_at")
+	private Timestamp createdAt;
 
-    @ManyToOne
-    @JoinColumn(name = "prv_affiliate_id", referencedColumnName = "aff_id", insertable = false, updatable = false)
-    private Affiliate affiliate;
+	@ManyToOne
+	@JoinColumn(name = "prv_affiliate_id", referencedColumnName = "aff_id", insertable = false, updatable = false)
+	private Affiliate affiliate;
 
-    @ManyToOne
-    @JoinColumn(name = "prv_transaction_id", referencedColumnName = "tra_id", insertable = false, updatable = false)
-    private Transaction transaction;
+	@ManyToOne
+	@JoinColumn(name = "prv_transaction_id", referencedColumnName = "tra_id", insertable = false, updatable = false)
+	private Transaction transaction;
 }
